refactor: Nested Call을 위한 메소드 시그니처 변경
