+++ src/main/java/com/weather/field/AbstractFieldGenerator.java
@@ -1,13 +1,15 @@
 package com.weather.field;
 
+import com.weather.JsonGenerator;
+
 import java.lang.reflect.Field;
 
 public abstract class AbstractFieldGenerator implements FieldGenerator{
 
     @Override
-    public String generateField(Field field) {
+    public String generateField(JsonGenerator jsonGenerator, Field field) {
         String key = createKey(field);
-        String value = createValue(field);
+        String value = createValue(jsonGenerator, field);
 
         return key + ":" + value;
     }
@@ -16,5 +18,5 @@ public abstract class AbstractFieldGenerator implements FieldGenerator{
         return "\"" + field.getName() + "\"";
     }
 
-    protected abstract String createValue(Field field);
+    protected abstract String createValue(JsonGenerator jsonGenerator, Field field);
 }
