+++ src/main/java/com/weather/field/FieldGenerator.java
@@ -1,10 +1,12 @@
 package com.weather.field;
 
+import com.weather.JsonGenerator;
+
 import java.lang.reflect.Field;
 
 public interface FieldGenerator {
 
-    String generateField(Field field);
+    String generateField(JsonGenerator jsonGenerator, Field field);
 
     boolean isSupported(Class<?> fieldType);
 }
