+++ src/main/java/com/weather/field/FieldGeneratorFactory.java
@@ -1,6 +1,9 @@
 package com.weather.field;
 
 
+import com.weather.StandardJsonGenerator;
+import com.weather.field.impl.*;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,6 +14,8 @@ public class FieldGeneratorFactory {
     private final static RealNumberGenerator REAL_NUMBER_GENERATOR = new RealNumberGenerator();
     private final static StringGenerator STRING_GENERATOR = new StringGenerator();
     private final static EnumGenerator ENUM_GENERATOR = new EnumGenerator();
+    private final static CollectionGenerator COLLECTION_GENERATOR = new CollectionGenerator();
+
 
     public static List<FieldGenerator> getFieldGenerators(){
         List<FieldGenerator> lists = new ArrayList<>();
@@ -19,6 +24,7 @@ public class FieldGeneratorFactory {
         lists.add(REAL_NUMBER_GENERATOR);
         lists.add(STRING_GENERATOR);
         lists.add(ENUM_GENERATOR);
+        lists.add(COLLECTION_GENERATOR);
         return lists;
     }
 }
