+++ src/main/java/com/weather/field/impl/CollectionGenerator.java
@@ -1,17 +1,20 @@
-package com.weather.field;
+package com.weather.field.impl;
+
+import com.weather.JsonGenerator;
+import com.weather.field.AbstractFieldGenerator;
 
 import java.lang.reflect.Field;
+import java.util.Collection;
 
-public class CollectionGenerator extends AbstractFieldGenerator{
+public class CollectionGenerator extends AbstractFieldGenerator {
 
     @Override
-    protected String createValue(Field field) {
+    protected String createValue(JsonGenerator jsonGenerator, Field field) {
         return null;
     }
 
     @Override
     public boolean isSupported(Class<?> fieldType) {
-        Collection.
-        return false;
+        return Collection.class.isAssignableFrom(fieldType);
     }
 }
