+++ src/main/java/com/weather/field/impl/IntegerGenerator.java
@@ -1,13 +1,16 @@
-package com.weather.field;
+package com.weather.field.impl;
+
+import com.weather.JsonGenerator;
+import com.weather.field.AbstractFieldGenerator;
 
 import java.lang.reflect.Field;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class IntegerGenerator extends AbstractFieldGenerator{
+public class IntegerGenerator extends AbstractFieldGenerator {
 
     @Override
-    protected String createValue(Field field) {
+    protected String createValue(JsonGenerator jsonGenerator, Field field) {
         return "1";
     }
