+++ src/main/java/com/weather/field/impl/RealNumberGenerator.java
@@ -1,10 +1,13 @@
-package com.weather.field;
+package com.weather.field.impl;
+
+import com.weather.JsonGenerator;
+import com.weather.field.AbstractFieldGenerator;
 
 import java.lang.reflect.Field;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
-public class RealNumberGenerator extends AbstractFieldGenerator{
+public class RealNumberGenerator extends AbstractFieldGenerator {
 
     @Override
     public boolean isSupported(Class<?> fieldType) {
@@ -15,7 +18,7 @@ public class RealNumberGenerator extends AbstractFieldGenerator{
     }
 
     @Override
-    protected String createValue(Field field) {
+    protected String createValue(JsonGenerator jsonGenerator, Field field) {
         return "1.0";
     }
 }
